Move and rename graphql query
