+++ metrics/github/api.py
@@ -4,8 +4,6 @@ import textwrap
 import requests
 import structlog
 
-from ..tools.dates import datetime_from_iso
-
 
 log = structlog.get_logger()
 
@@ -84,43 +82,3 @@ class GitHubClient:
             yield from page["nodes"]
             more_pages = page["pageInfo"]["hasNextPage"]
             cursor = page["pageInfo"]["endCursor"]
-
-
-def iter_repo_prs(client, repo):
-    query = """
-    query prs($cursor: String, $org: String!, $repo: String!) {
-      organization(login: $org) {
-        repository(name: $repo) {
-          pullRequests(first: 100, after: $cursor) {
-            nodes {
-              author {
-                login
-              }
-              number
-              createdAt
-              closedAt
-              mergedAt
-            }
-            pageInfo {
-              endCursor
-              hasNextPage
-            }
-          }
-        }
-      }
-    }
-    """
-    for pr in client.get_query(
-        query,
-        path=["organization", "repository", "pullRequests"],
-        repo=repo["name"],
-    ):
-        yield {
-            "org": client.org,
-            "repo": repo["name"],
-            "repo_archived_at": datetime_from_iso(repo["archivedAt"]),
-            "author": pr["author"]["login"],
-            "closed_at": datetime_from_iso(pr["closedAt"]),
-            "created_at": datetime_from_iso(pr["createdAt"]),
-            "merged_at": datetime_from_iso(pr["mergedAt"]),
-        }
