+++ metrics/github/query.py
@@ -1,3 +1,6 @@
+from metrics.tools.dates import datetime_from_iso
+
+
 def repos(client):
     query = """
     query repos($cursor: String, $org: String!) {
@@ -46,3 +49,43 @@ def vulnerabilities(client, repo):
         org=client.org,
         repo=repo["name"],
     )
+
+
+def prs(client, repo):
+    query = """
+    query prs($cursor: String, $org: String!, $repo: String!) {
+      organization(login: $org) {
+        repository(name: $repo) {
+          pullRequests(first: 100, after: $cursor) {
+            nodes {
+              author {
+                login
+              }
+              number
+              createdAt
+              closedAt
+              mergedAt
+            }
+            pageInfo {
+              endCursor
+              hasNextPage
+            }
+          }
+        }
+      }
+    }
+    """
+    for pr in client.get_query(
+        query,
+        path=["organization", "repository", "pullRequests"],
+        repo=repo["name"],
+    ):
+        yield {
+            "org": client.org,
+            "repo": repo["name"],
+            "repo_archived_at": datetime_from_iso(repo["archivedAt"]),
+            "author": pr["author"]["login"],
+            "closed_at": datetime_from_iso(pr["closedAt"]),
+            "created_at": datetime_from_iso(pr["createdAt"]),
+            "merged_at": datetime_from_iso(pr["mergedAt"]),
+        }
